Refactor: move methods to element.cpp
